refactor: move utility functions for callsign encoding, decoding, and random generation; move TUI for data display; move client base functions
